Impulse-Nutrition/gorgias_mcp/pyproject.toml
[project]
name = "gorgias-mcp"
version = "0.2.0"
description = "MCP server for Gorgias customer support (view-based filtering)"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.9.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "responses>=0.24",
]

[project.scripts]
gorgias-mcp = "gorgias_mcp.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/gorgias_mcp"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
